+++ packages/metagame/src/entry_requirement/tests/mocks/rejecting_entry_validator_mock.cairo
@@ -70,12 +70,12 @@ pub mod RejectingEntryValidatorMock {
             context_id: u64,
             player_address: ContractAddress,
             qualification: Span<felt252>,
-        ) -> Option<u8> {
+        ) -> Option<u32> {
             Option::Some(0)
         }
 
         fn add_config(
-            ref self: ContractState, context_id: u64, entry_limit: u8, config: Span<felt252>,
+            ref self: ContractState, context_id: u64, entry_limit: u32, config: Span<felt252>,
         ) {}
 
         fn add_entry(
